refactor(WinningRank): Move WinningRank enum to domain package
