+++ src/main/kotlin/lotto/domain/WinningStats.kt
@@ -1,7 +1,5 @@
 package lotto.domain
 
-import lotto.WinningRank
-
 data class WinningStats(
     val stats: Map<WinningRank, Int>
 )
